refactor(sessions): split SQLite session accessor
