+++ src/config/sessions/session-accessor.sqlite-session-row.ts
@@ -1,0 +1,103 @@
+import {
+  normalizeSqliteChatType,
+  normalizeSqliteText,
+} from "./session-accessor.sqlite-normalize.js";
+import { bindSessionEntryProvenance } from "./session-accessor.sqlite-provenance.js";
+import { normalizeSqliteStatus } from "./session-accessor.sqlite-status.js";
+import type { SessionEntry } from "./types.js";
+
+export function normalizeSqliteSessionEntryTimestamp(entry: SessionEntry): SessionEntry {
+  if (typeof entry.updatedAt === "number" && Number.isFinite(entry.updatedAt)) {
+    return entry;
+  }
+  const updatedAt =
+    typeof entry.sessionStartedAt === "number" && Number.isFinite(entry.sessionStartedAt)
+      ? entry.sessionStartedAt
+      : Date.now();
+  return { ...entry, updatedAt };
+}
+
+export function bindSqliteSessionRoot(params: {
+  entry: SessionEntry;
+  sessionKey: string;
+  updatedAt: number;
+}) {
+  const updatedAt = Number.isFinite(params.entry.updatedAt)
+    ? params.entry.updatedAt
+    : params.updatedAt;
+  return {
+    session_id: params.entry.sessionId,
+    session_key: params.sessionKey,
+    session_scope: resolveSqliteSessionScope(params.entry, params.sessionKey),
+    created_at: resolveSqliteSessionCreatedAt(params.entry, updatedAt),
+    updated_at: updatedAt,
+    ...bindSessionEntryProvenance(params.entry),
+    started_at: finiteSqliteNumber(params.entry.startedAt),
+    ended_at: finiteSqliteNumber(params.entry.endedAt),
+    status: normalizeSqliteStatus(params.entry.status),
+    chat_type: normalizeSqliteChatType(params.entry.chatType),
+    channel: resolveSqliteSessionChannel(params.entry),
+    account_id: resolveSqliteSessionAccountId(params.entry),
+    primary_conversation_id: null,
+    model_provider: normalizeSqliteText(params.entry.modelProvider),
+    model: normalizeSqliteText(params.entry.model),
+    agent_harness_id: normalizeSqliteText(params.entry.agentHarnessId),
+    parent_session_key: normalizeSqliteText(params.entry.parentSessionKey),
+    spawned_by: normalizeSqliteText(params.entry.spawnedBy),
+    display_name: resolveSqliteSessionDisplayName(params.entry),
+  };
+}
+
+function resolveSqliteSessionScope(
+  entry: Pick<SessionEntry, "chatType">,
+  sessionKey: string,
+): "conversation" | "shared-main" | "group" | "channel" {
+  const chatType = normalizeSqliteChatType(entry.chatType);
+  const normalizedKey = sessionKey.trim().toLowerCase();
+  if (chatType === "direct" && (normalizedKey === "main" || normalizedKey.endsWith(":main"))) {
+    return "shared-main";
+  }
+  if (chatType === "group" || chatType === "channel") {
+    return chatType;
+  }
+  return "conversation";
+}
+
+function resolveSqliteSessionCreatedAt(entry: SessionEntry, updatedAt: number): number {
+  for (const candidate of [entry.sessionStartedAt, entry.startedAt, entry.updatedAt, updatedAt]) {
+    if (typeof candidate === "number" && Number.isFinite(candidate) && candidate >= 0) {
+      return candidate;
+    }
+  }
+  return updatedAt;
+}
+
+function finiteSqliteNumber(value: unknown): number | null {
+  return typeof value === "number" && Number.isFinite(value) ? value : null;
+}
+
+function resolveSqliteSessionChannel(entry: SessionEntry): string | null {
+  return (
+    normalizeSqliteText(entry.channel) ??
+    normalizeSqliteText(entry.deliveryContext?.channel) ??
+    normalizeSqliteText(entry.lastChannel) ??
+    normalizeSqliteText(entry.origin?.provider)
+  );
+}
+
+function resolveSqliteSessionAccountId(entry: SessionEntry): string | null {
+  return (
+    normalizeSqliteText(entry.deliveryContext?.accountId) ??
+    normalizeSqliteText(entry.lastAccountId) ??
+    normalizeSqliteText(entry.origin?.accountId)
+  );
+}
+
+function resolveSqliteSessionDisplayName(entry: SessionEntry): string | null {
+  return (
+    normalizeSqliteText(entry.displayName) ??
+    normalizeSqliteText(entry.label) ??
+    normalizeSqliteText(entry.subject) ??
+    normalizeSqliteText(entry.groupId)
+  );
+}
